+++ HISTORY.md
@@ -6,6 +6,7 @@
 
 ### 2026-07-27
 * symbols: refactored drawing a single 8-spike gear symbol
+* symbols: added a single "gear" symbol, changed "gears" to a pair of interlocking gears 
 
 ### 2026-07-24
 * new dev version